refactor: separated embedded objects
