+++ src/main/java/com/translink/api/repository/model/Route.java
@@ -1,5 +1,6 @@
 package com.translink.api.repository.model;
 
+import com.translink.api.repository.model.embed.RouteType;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.validator.constraints.URL;
