--- src/main/java/com/translink/api/repository/model/StopType.java
@@ -1,4 +1,0 @@
-package com.translink.api.repository.model;
-
-public enum StopType {
-}
