+++ src/main/java/com/translink/api/repository/model/Trip.java
@@ -1,5 +1,8 @@
 package com.translink.api.repository.model;
 
+import com.translink.api.repository.model.embed.Direction;
+import com.translink.api.repository.model.embed.Shape;
+import com.translink.api.repository.model.embed.StopTimes;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
