+++ src/main/java/com/translink/api/repository/model/embed/Direction.java
@@ -1,4 +1,4 @@
-package com.translink.api.repository.model;
+package com.translink.api.repository.model.embed;
 
 public enum Direction {
     OUTBOUND, INBOUND
