+++ src/main/java/com/translink/api/repository/model/embed/ExceptionType.java
@@ -1,4 +1,4 @@
-package com.translink.api.repository.model;
+package com.translink.api.repository.model.embed;
 
 public enum ExceptionType {
     UNKNOWN, ADDED, REMOVED
