+++ src/main/java/com/translink/api/repository/model/embed/RouteType.java
@@ -1,4 +1,4 @@
-package com.translink.api.repository.model;
+package com.translink.api.repository.model.embed;
 
 public enum RouteType {
     LIGHT_RAIL, SUBWAY, RAIL, BUS, FERRY, UNKNOWN
