+++ src/main/java/com/translink/api/repository/model/embed/StopDropOffType.java
@@ -1,4 +1,4 @@
-package com.translink.api.repository.model;
+package com.translink.api.repository.model.embed;
 
 public enum StopDropOffType {
     REGULAR, PHONE_ARRANGED, DRIVER_ARRANGED
