+++ src/main/java/com/translink/api/repository/model/embed/StopPickupType.java
@@ -1,4 +1,4 @@
-package com.translink.api.repository.model;
+package com.translink.api.repository.model.embed;
 
 public enum StopPickupType {
     REGULAR, NOT_AVAILABLE, PHONE_ARRANGED, DRIVER_ARRANGED
