+++ src/main/java/com/translink/api/repository/model/embed/StopType.java
@@ -1,0 +1,4 @@
+package com.translink.api.repository.model.embed;
+
+public enum StopType {
+}
